Fix CI and address PR review: auction resume, validation split, engine hygiene
- CI: biome format/imports on gameStateMachine, validation index, optional effects
- Leveraged buyout uses waiting_for_roll resume phase on round-start draws
- Shared marketEventPrimitives; seeded algorithmic flash trade rolls
- gameReducer delegates via deepClone; map game.* throws to GameEngineErrorKey
- Handshake sign idempotency; HandshakePhasePanel extraction
- gameStateView unified client state builder; split validation into gameSchemas
- Tests for auction resumePhase and repeat handshake sign
